added Contact class used super(...,isPrivate) as constructors
